requirements.txt
asgiref==3.4.1
certifi==2020.6.20
charset-normalizer==2.0.9
Django==4.0
djangorestframework==3.13.1
idna==3.3
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
